requirements.txt
#Core LangChain and Hugging Face packages

langchain-huggingface
transformers
huggingface-hub

#PyTorch is required by transformers

torch
#Accelerate is needed for device_map="auto" (to use GPU)

accelerate

#ptional, but good to have from your original list

python-dotenv
numpy